React Project Configuration for Github Pages and Using Build as Static files
