+++ components.html
@@ -39,7 +39,7 @@
     </nav>
 
     <div class="wrapper" id="component-items">
-        
+
         <!-- components will be fetched here -->
 
     </div>
