--- components/react-components/src/App.js
@@ -1,19 +1,0 @@
-import { BrowserRouter, Routes, Route } from "react-router-dom";
-import "./App.css";
-import Layout from "./pages/layout/Layout";
-import Slider from "./pages/slider/Slider";
-import Error from "./pages/error/Error";
-function App() {
-  return (
-    <BrowserRouter>
-      {/* <Navbar /> */}
-      <Routes>
-        <Route path="/" element={<Layout />} />
-        <Route path="/slider" element={<Slider />} />
-        <Route path="*" element={<Error />} />
-      </Routes>
-    </BrowserRouter>
-  );
-}
-
-export default App;
