--- components/react-components/src/index.js
@@ -1,33 +1,0 @@
-// import ReactDOM from "react-dom/client";
-// import { BrowserRouter, Routes, Route } from "react-router-dom";
-// import Slider from "./pages/slider/Slider";
-// import Error from "./pages/error/Error";
-// import Layout from "./pages/layout/Layout";
-
-// export default function App() {
-//   return (
-//     <BrowserRouter>
-//       <Routes>
-//         <Route path="/" element={<Layout />}>
-//           <Route path="/slider" element={<Slider />} />
-//           {/* <Route path="contact" element={<Contact />} /> */}
-//           <Route path="*" element={<Error />} />
-//         </Route>
-//       </Routes>
-//     </BrowserRouter>
-//   );
-// }
-
-// const root = ReactDOM.createRoot(document.getElementById("root"));
-// root.render(<App />);
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import './index.css';
-import App from './App';
-
-const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>
-);
