+++ components/react/package.json
@@ -2,6 +2,7 @@
   "name": "react-components",
   "version": "0.1.0",
   "private": true,
+  "homepage": "http://127.0.0.1:5500/components/react/build/",
   "dependencies": {
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
