+++ components/react/src/pages/slider/Slider.jsx
