+++ components/react/src/pages/slider/slider.css
